// src/types/sessions/getExcelSessionsData.types.ts
export type GetSessions_ExcelData_Api_Options = {
  pagelimit?: string;
  page?: string;
  enrollment_id?: string;
  tutor_id?: string;
  subject_id?: string;
  student_ids?: string;
  curriculum_id?: string;
  grade_id?: string;
  board_id?: string;
  conclusion_type?: string;
  start_time?: string;
  end_time?: string;
};